Add punch king policies and fix interstitial rewards

+++ src/main/java/com/money_hunter/application/dto/response/WeeklyPunchKingResponse.java
@@ -12,6 +12,6 @@ public record WeeklyPunchKingResponse(
 		int ultimateCooldownSeconds,
 		long goldRewardDivisor,
 		long nextGoldRewardScore,
-		int nextSkillPointRewardScore
+		long nextSkillPointRewardScore
 ) {
 }
